requirements.txt
# GP MLOps Platform - Complete Requirements

# Core Web Framework
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1

# Database
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.9

# Data Processing
pandas==2.2.3
numpy==1.26.4

# Machine Learning
scikit-learn==1.5.2
joblib==1.4.2
torch==2.2.0

# API Support
python-multipart==0.0.17
aiohttp==3.11.10
httpx==0.27.2

# Caching and Queue
redis==5.2.1
celery==5.4.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Monitoring
prometheus-client==0.20.0

# AWS Integration
boto3==1.35.75
botocore==1.35.75

# Additional packages for full functionality  
snowflake-connector-python==3.17.1
snowflake-sqlalchemy==1.6.1
mlflow==2.19.0

# Computer Vision packages
opencv-python==4.10.0.84
Pillow==11.1.0
torchvision==0.17.0

# Additional FastAPI/API packages  
requests==2.32.3